+++ crucible-mir/src/Mir/Mir.hs
@@ -91,11 +91,7 @@ data Ty =
       | TyAdt !DefId !DefId !Substs -- first DefId is the monomorphized name, second is pre-mono
       | TyFnDef !DefId
       | TyClosure [Ty]      -- the Tys are the types of the upvars
-      | TyCoroutine
-        -- ^ @crucible-mir@ does not support coroutines (#1369), but we
-        -- nevertheless include this as a 'Ty' so that we can successfully
-        -- translate code that mentions it. Provided that that code is never
-        -- simulated, this should work out.
+      | TyCoroutine CoroutineArgs
       | TyCoroutineClosure [Ty]     -- the Tys are the types of the upvars
       | TyStr
       | TyFnPtr !FnSig              -- written as fn() -> i32
@@ -125,6 +121,24 @@ data Ty =
       -- which runs just after JSON decoding is done.
       | TyInterned TyName
       deriving (Eq, Ord, Show, Generic)
+
+-- | Details about a coroutine type.  See Note [coroutine representation] in
+-- @Mir.TransTy@ for how these correspond to the `TypeRepr` for the coroutine
+-- type.
+data CoroutineArgs = CoroutineArgs
+  { -- | The discriminant type for this coroutine.  This is often `u32`, but
+    -- can theoretically vary.
+    _caDiscrTy :: !Ty
+    -- | The types of the upvar fields.
+  , _caUpvarTys :: ![Ty]
+    -- | The types of the fields for storing saved locals.
+  , _caSavedTys :: ![Ty]
+    -- | Maps (variant index, field index) to an index into `_caSavedTys`.
+    -- This is used to determine which field is accessed by downcast field
+    -- projections, e.g. `(co as variant#1).2`.
+  , _caFieldMap :: !(Map (Int, Int) Int)
+  }
+  deriving (Eq, Ord, Show, Generic)
 
 data NamedTy = NamedTy
   { _ntName :: Text
@@ -633,7 +647,7 @@ data AggregateKind =
         AKArray Ty
       | AKTuple
       | AKClosure
-      | AKCoroutine
+      | AKCoroutine CoroutineArgs
       | AKCoroutineClosure
       | AKRawPtr Ty Mutability
       deriving (Show,Eq, Ord, Generic)
@@ -700,6 +714,7 @@ makeLenses ''Layout
 makeLenses ''LangItem
 makeLenses ''Statement
 makeLenses ''Terminator
+makeLenses ''CoroutineArgs
 makeWrapped ''Substs
 
 --------------------------------------------------------------------------------------
@@ -848,7 +863,7 @@ instance TypeOf Rvalue where
   typeOf (Aggregate AKTuple ops) = TyTuple $ map typeOf ops
   typeOf (Aggregate AKClosure ops) = TyClosure $ map typeOf ops
   typeOf (Aggregate (AKRawPtr ty mutbl) _ops) = TyRawPtr ty mutbl
-  typeOf (Aggregate AKCoroutine _ops) = TyCoroutine
+  typeOf (Aggregate (AKCoroutine ca) _ops) = TyCoroutine ca
   typeOf (Aggregate AKCoroutineClosure ops) = TyCoroutineClosure $ map typeOf ops
   typeOf (RAdtAg (AdtAg _ _ _ ty _)) = ty
   typeOf (ShallowInitBox _ ty) = ty